+++ packages/bsky/src/api/app/bsky/graph/getLists.ts
@@ -1,12 +1,13 @@
 import { mapDefined } from '@atproto/common'
 import { Server } from '../../../../lexicon'
 import { QueryParams } from '../../../../lexicon/types/app/bsky/graph/getLists'
+import { REFERENCELIST } from '../../../../lexicon/types/app/bsky/graph/defs'
 import AppContext from '../../../../context'
 import {
   createPipeline,
   HydrationFnInput,
-  noRules,
   PresentationFnInput,
+  RulesFnInput,
   SkeletonFnInput,
 } from '../../../../pipeline'
 import { HydrateCtx, Hydrator } from '../../../../hydration/hydrator'
@@ -14,7 +15,12 @@ import { Views } from '../../../../views'
 import { clearlyBadCursor, resHeaders } from '../../../util'
 
 export default function (server: Server, ctx: AppContext) {
-  const getLists = createPipeline(skeleton, hydration, noRules, presentation)
+  const getLists = createPipeline(
+    skeleton,
+    hydration,
+    noReferenceLists,
+    presentation,
+  )
   server.app.bsky.graph.getLists({
     auth: ctx.authVerifier.optionalStandardOrRole,
     handler: async ({ params, auth, req }) => {
@@ -57,6 +63,17 @@ const hydration = async (
   const { ctx, params, skeleton } = input
   const { listUris } = skeleton
   return ctx.hydrator.hydrateLists(listUris, params.hydrateCtx)
+}
+
+const noReferenceLists = (
+  input: RulesFnInput<Context, Params, SkeletonState>,
+) => {
+  const { skeleton, hydration } = input
+  skeleton.listUris = skeleton.listUris.filter((uri) => {
+    const list = hydration.lists?.get(uri)
+    return list?.record.purpose !== REFERENCELIST
+  })
+  return skeleton
 }
 
 const presentation = (